src/main/java/wooteco/subway/domain/Line.java
package wooteco.subway.domain;

import java.util.List;
import java.util.Objects;
import wooteco.subway.exception.ExceptionMessage;
import wooteco.subway.exception.DomainException;

public class Line {

    private final Long id;
    private final String name;
    private final String color;
    private final Long extraFare;

    public Line(Long id, String name, String color, Long extraFare) {
        if (name.isBlank()) {
            throw new DomainException(ExceptionMessage.BLANK_LINE_NAME.getContent());
        }
        this.id = id;
        this.name = name;
        this.color = color;
        this.extraFare = extraFare;
    }

    public Line(Long id, String name, String color) {
        this(id, name, color, 0L);
    }

    public static Line withoutIdOf(String name, String color, Long extraFare) {
        return new Line(null, name, color, extraFare);
    }

    public Long getId() {
        return id;
    }

    public String getName() {
        return name;
    }

    public String getColor() {
        return color;
    }

    public Long getExtraFare() {
        return extraFare;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) {
            return true;
        }
        if (o == null || getClass() != o.getClass()) {
            return false;
        }
        Line line = (Line) o;
        return Objects.equals(id, line.id);
    }

    @Override
    public int hashCode() {
        return Objects.hash(id);
    }
}
